new versions of files with basic SQL functionality
After much trouble (I bravely but foolishly had chosen to use Cygwin as my terminal app, the course has no tutorial for it and I had to Google a LOT of arcane errors) I got SQL working. Here is the example app in the course material, adjusted to work on my computer.